+++ src/Firebase/Exception/RemoteConfig/VersionNotFound.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kreait\Firebase\Exception\RemoteConfig;
+
+use Kreait\Firebase\Exception\RemoteConfigException;
+use Kreait\Firebase\RemoteConfig\VersionNumber;
+
+final class VersionNotFound extends RemoteConfigException
+{
+    public static function withVersionNumber(VersionNumber $versionNumber): self
+    {
+        return new self('Version #'.$versionNumber.' could not be found.');
+    }
+}