+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Features
 
-* Added support for Remote Config Template validation ([Documentation](https://firebase-php.readthedocs.io/en/latest/remote-config.html#validation)) 
+* Added support for Remote Config Template validation ([Documentation](https://firebase-php.readthedocs.io/en/latest/remote-config.html#validation))
+* Added support for working with the Remote Config History ([Documentation](https://firebase-php.readthedocs.io/en/latest/remote-config.html#change-history))
 
 ## 4.15.1 - 2018-08-07